Extract @stlite/app-packager from @stlite/desktop's bin
The dump-stlite-desktop-artifacts orchestration (vendor pyodide,
snapshot site-packages, copy app files) is generic — it just happens
to live inside desktop's bin today. Pull it out so the upcoming
@stlite/cli's `web` and `desktop` commands can call the same code path
without depending on @stlite/desktop's heavier surface.

- packages/app-packager: new library exposing `packageApp(opts)` plus
  `Logger`/`consoleLogger`, `DEFAULT_PYODIDE_SOURCE`, helpers
  `readRequirementsTxt`/`normalizePyodideSource`. Pyodide-package
  reader moves here too.
- packageApp is restructured as additive — it writes app_files/,
  pyodide/, site-packages-snapshot.tar.gz, prebuilt-packages.txt into
  destDir but does NOT touch the SPA bundle. Caller manages destDir
  lifecycle and copies its own shell (desktop's CRA SPA, or the
  upcoming `stlite web`'s minimal index.html). This separation lets
  one packager primitive serve both targets.
- packages/desktop: bin-src/dump_artifacts shrinks to a thin
  orchestrator that reads package.json#stlite.desktop, copies its own
  CRA SPA into destDir, calls packageApp, then writes
  stlite-manifest.json. dumpManifest + coerceDesktopAppManifest move
  into electron/manifest.ts (alongside the schema and readManifest)
  so the file is the single source of truth for the desktop manifest;
  the spec moves with the code. build_bin.js externalizes
  @stlite/app-packager (bundle 280kb -> 224kb).

The bin's CLI surface is unchanged.
