+++ packages/desktop/bin-src/dump_artifacts/index.ts
@@ -7,15 +7,15 @@ import fsPromises from "node:fs/promises";
 import { fileURLToPath } from "node:url";
 import fsExtra from "fs-extra";
 import {
-  loadPyodide,
-  version as pyodideVersion,
-  type PyodideInterface,
-} from "pyodide";
-import { PrebuiltPackagesDataReader } from "./pyodide_packages";
-import { dumpManifest } from "./manifest";
+  packageApp,
+  readRequirementsTxt,
+  DEFAULT_PYODIDE_SOURCE,
+  normalizePyodideSource,
+  type Logger,
+} from "@stlite/app-packager";
+import { validateRequirements } from "@stlite/common";
+import { dumpManifest } from "../../electron/manifest";
 import { readConfig } from "./config";
-import { validateRequirements, parseRequirementsTxt } from "@stlite/common";
-import { glob } from "glob";
 import { logger } from "./logger";
 
 const pathFromScriptToBuild =
@@ -38,282 +38,53 @@ const stliteLibWheelFileName =
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
-/**
- * Ensure that the given package is loaded by throwing an error if any error occurs in the loadPackage() function.
- */
-async function ensureLoadPackage(
-  pyodide: PyodideInterface,
-  packageName: string | string[],
-) {
-  const errorMessages: string[] = [];
-  const errorCallback = (message: string): void => {
-    errorMessages.push(message);
-  };
-  await pyodide.loadPackage(packageName, { errorCallback });
-  if (errorMessages.length > 0) {
-    throw new Error(errorMessages.join("\n"));
-  }
-}
+const winstonAdapter: Logger = {
+  debug: (msg, ...meta) => logger.debug(msg, ...meta),
+  info: (msg, ...meta) => logger.info(msg, ...meta),
+  warn: (msg, ...meta) => logger.warn(msg, ...meta),
+  error: (msg, ...meta) => logger.error(msg, ...meta),
+};
 
-interface CopyBuildDirectoryOptions {
+interface CopyShellDirectoryOptions {
+  sourceDir: string;
+  destDir: string;
   keepOld: boolean;
-  copyTo: string;
 }
-async function copyBuildDirectory(options: CopyBuildDirectoryOptions) {
+async function copyShellDirectory(options: CopyShellDirectoryOptions) {
   logger.info(
-    "Copy the build directory (the bare built app files) to this directory...",
+    "Copy the desktop SPA shell (the bare built app files) to the destination directory...",
   );
 
-  const sourceDir = path.resolve(__dirname, pathFromScriptToBuild);
-  const sourceDirStat = await fsPromises.stat(sourceDir);
+  const sourceDirStat = await fsPromises.stat(options.sourceDir);
   if (!sourceDirStat.isDirectory()) {
-    throw new Error(`The source ${sourceDir} does not exist.`);
+    throw new Error(`The source ${options.sourceDir} does not exist.`);
   }
 
-  if (sourceDir === options.copyTo) {
+  if (options.sourceDir === options.destDir) {
     logger.warn(
-      `sourceDir == destDir (${sourceDir}). Are you in the development environment? Skip copying the directory.`,
+      `sourceDir == destDir (${options.sourceDir}). Are you in the development environment? Skip copying the directory.`,
     );
     return;
   }
 
   if (options.keepOld) {
     try {
-      await fsPromises.access(options.copyTo);
-      logger.info(`${options.copyTo} already exists. Use it and skip copying.`);
+      await fsPromises.access(options.destDir);
+      logger.info(
+        `${options.destDir} already exists. Use it and skip copying.`,
+      );
       return;
     } catch {
-      // If the destination directory does not exist
       throw new Error(
-        `${options.copyTo} does not exist even though the \`keepOld\` option is specified`,
+        `${options.destDir} does not exist even though the \`keepOld\` option is specified`,
       );
     }
   }
 
-  logger.debug(`Copy ${sourceDir} to ${options.copyTo}`);
-  await fsPromises.rm(options.copyTo, { recursive: true, force: true });
-  await fsExtra.copy(sourceDir, options.copyTo, { errorOnExist: true });
-}
-
-interface LoadUsedPrebuiltPackagesOptions {
-  pyodideSource: string;
-  pyodideRuntimeDir: string;
-  requirements: string[];
-}
-/**
- * Load the Pyodide runtime and install the given requirements to load the prebuilt packages used by the requirements.
- * Those prebuilt package wheels will be downloaded/copied to a local directory, `pyodideRuntimeDir`.
- * Pyodide's caching mechanism available in the Node environment is used here as the wheel file downloader.
- * `pyodideRuntimeDir` should be "build/pyodide" so that the downloaded/copied files will be vendored in the app executable.
- * This vendoring and runtime-loading mechanism is necessary to avoid problems such as https://github.com/whitphx/stlite/issues/558
- */
-async function saveUsedPrebuiltPackages(
-  options: LoadUsedPrebuiltPackagesOptions,
-): Promise<string[]> {
-  if (options.requirements.length === 0) {
-    return [];
-  }
-
-  const pyodide = await loadPyodide({
-    packageBaseUrl: options.pyodideSource,
-    packageCacheDir: options.pyodideRuntimeDir,
-  });
-
-  await installPackages(pyodide, {
-    requirements: options.requirements,
-  });
-
-  return Object.entries(pyodide.loadedPackages)
-    .filter(([, channel]) => channel === "default channel")
-    .map(([name]) => name);
-}
-
-async function prepareLocalWheel(
-  pyodide: PyodideInterface,
-  localPath: string,
-): Promise<string> {
-  logger.debug(`Preparing the local wheel: %s`, localPath);
-
-  const data = await fsPromises.readFile(localPath);
-  const emfsPath = "/tmp/" + path.basename(localPath);
-  pyodide.FS.writeFile(emfsPath, data);
-
-  const requirement = `emfs:${emfsPath}`;
-  logger.debug(`The local wheel %s is prepared as %s`, localPath, requirement);
-  return requirement;
-}
-
-interface InstallStreamlitWheelsOptions {
-  requirements: string[];
-}
-async function installPackages(
-  pyodide: PyodideInterface,
-  options: InstallStreamlitWheelsOptions,
-) {
-  await ensureLoadPackage(pyodide, "micropip");
-  const micropip = pyodide.pyimport("micropip");
-
-  const requirements: string[] = [...options.requirements];
-
-  const wheelsDir = path.join(__dirname, pathFromScriptToWheels);
-  const stliteLibWheel = await prepareLocalWheel(
-    pyodide,
-    path.join(wheelsDir, stliteLibWheelFileName),
-  );
-  requirements.push(stliteLibWheel);
-  const streamlitWheel = await prepareLocalWheel(
-    pyodide,
-    path.join(wheelsDir, streamlitWheelFileName),
-  );
-  requirements.push(streamlitWheel);
-
-  logger.info("Install the packages: %j", requirements);
-  await micropip.install.callKwargs(requirements, { keep_going: true });
-}
-
-interface CreateSitePackagesSnapshotOptions {
-  requirements: string[];
-  usedPrebuiltPackages: string[];
-  pyodideRuntimeDir: string;
-  pyodideSource: string;
-  saveTo: string;
-}
-async function createSitePackagesSnapshot(
-  options: CreateSitePackagesSnapshotOptions,
-) {
-  logger.info("Create the site-packages snapshot file...");
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const pyodide: PyodideInterface & { FS: any } = // XXX: `{ FS: any }` is a temporary workaround to fix the type error.
-    await loadPyodide({
-      packageBaseUrl: options.pyodideSource,
-      packageCacheDir: options.pyodideRuntimeDir,
-    });
-
-  await ensureLoadPackage(pyodide, "micropip");
-  const micropip = pyodide.pyimport("micropip");
-
-  const prebuiltPackagesDataReader = new PrebuiltPackagesDataReader(
-    options.pyodideSource,
-  );
-
-  const mockedPackages: string[] = [];
-  if (options.usedPrebuiltPackages.length > 0) {
-    logger.info(
-      "Mocking prebuilt packages so that they will not be included in the site-packages snapshot because these will be installed from the vendored wheel files at runtime...",
-    );
-    for (const pkg of options.usedPrebuiltPackages) {
-      const packageInfo =
-        await prebuiltPackagesDataReader.getPackageInfoByName(pkg);
-      if (packageInfo == null) {
-        throw new Error(`Package ${pkg} is not found in the lock file.`);
-      }
-
-      logger.debug(`Mock ${packageInfo.name} ${packageInfo.version}`);
-      micropip.add_mock_package(packageInfo.name, packageInfo.version);
-      mockedPackages.push(packageInfo.name);
-    }
-  }
-
-  logger.info(`Install the requirements %j`, options.requirements);
-
-  await installPackages(pyodide, {
-    requirements: options.requirements,
-  });
-
-  logger.info("Remove the mocked packages: %j", mockedPackages);
-  mockedPackages.forEach((pkg) => micropip.remove_mock_package(pkg));
-
-  logger.info("Archive the site-packages director(y|ies)");
-  const archiveFilePath = "/tmp/site-packages-snapshot.tar.gz";
-  await pyodide.runPythonAsync(`
-    import os
-    import tarfile
-    import site
-
-    site_packages_dirs = site.getsitepackages()
-
-    tar_file_name = '${archiveFilePath}'
-    with tarfile.open(tar_file_name, mode='w:gz') as gzf:
-        for site_packages in site_packages_dirs:
-            print("Add site-package:", site_packages)
-            print(os.listdir(site_packages))
-            gzf.add(site_packages)
-  `);
-
-  logger.info("Extract the archive file from EMFS");
-  const archiveBin = pyodide.FS.readFile(archiveFilePath);
-
-  logger.info(`Save the archive file (${options.saveTo})`);
-  await fsPromises.writeFile(options.saveTo, archiveBin);
-}
-
-interface CopyAppDirectoryOptions {
-  cwd: string;
-  filePathPatterns: string[];
-  destAppDir: string;
-}
-
-async function copyAppDirectory(options: CopyAppDirectoryOptions) {
-  logger.info("Copy the app directory...");
-
-  await Promise.all(
-    options.filePathPatterns.map(async (pattern) => {
-      const fileRelPaths = await glob(pattern, {
-        cwd: options.cwd,
-        absolute: false,
-      });
-
-      if (fileRelPaths.length === 0) {
-        logger.warn(
-          `No files match the pattern "${pattern}" in "${options.cwd}".`,
-        );
-        return;
-      }
-
-      await Promise.all(
-        fileRelPaths.map(async (relPath) => {
-          const srcPath = path.resolve(options.cwd, relPath);
-          const destPath = path.resolve(options.destAppDir, relPath);
-          logger.debug(`Copy ${srcPath} to ${destPath}`);
-          await fsExtra.copy(srcPath, destPath, {
-            errorOnExist: true,
-          });
-        }),
-      );
-    }),
-  );
-}
-
-async function assertAppDirContainsEntrypoint(
-  appDir: string,
-  entrypoint: string,
-) {
-  try {
-    await fsPromises.access(path.resolve(appDir, entrypoint));
-  } catch {
-    throw new Error(
-      `The entrypoint file "${entrypoint}" is not included in the bundled files.`,
-    );
-  }
-}
-
-async function readRequirements(
-  requirementsTxtPath: string,
-): Promise<string[]> {
-  const requirementsTxtData = await fsPromises.readFile(requirementsTxtPath, {
-    encoding: "utf-8",
-  });
-  return parseRequirementsTxt(requirementsTxtData);
-}
-
-async function writePrebuiltPackagesTxt(
-  prebuiltPackagesTxtPath: string,
-  prebuiltPackages: string[],
-): Promise<void> {
-  const prebuiltPackagesTxtData = prebuiltPackages.join("\n");
-  await fsPromises.writeFile(prebuiltPackagesTxtPath, prebuiltPackagesTxtData, {
-    encoding: "utf-8",
+  logger.debug(`Copy ${options.sourceDir} to ${options.destDir}`);
+  await fsPromises.rm(options.destDir, { recursive: true, force: true });
+  await fsExtra.copy(options.sourceDir, options.destDir, {
+    errorOnExist: true,
   });
 }
 
@@ -358,13 +129,8 @@ yargs(hideBin(process.argv))
     type: "string",
     describe:
       "The base URL or path of the Pyodide files to download or copy, such as the prebuild package wheels and pyodide-lock.json",
-    default: `https://cdn.jsdelivr.net/pyodide/v${pyodideVersion}/full/`,
-    coerce: (urlOrPath) => {
-      if (!urlOrPath.endsWith("/")) {
-        urlOrPath += "/";
-      }
-      return urlOrPath;
-    },
+    default: DEFAULT_PYODIDE_SOURCE,
+    coerce: normalizePyodideSource,
   })
   .options("logLevel", {
     type: "string",
@@ -375,7 +141,6 @@ yargs(hideBin(process.argv))
   .parseAsync()
   .then(async (args) => {
     logger.level = args.logLevel;
-
     logger.debug(`Command line arguments: %j`, args);
 
     const projectDir = args.project;
@@ -401,11 +166,11 @@ yargs(hideBin(process.argv))
       config.requirementsTxtFilePaths,
     );
 
-    const dependenciesFromRequirementsTxt = await Promise.all(
-      config.requirementsTxtFilePaths.map(async (requirementsTxtPath) => {
-        return readRequirements(requirementsTxtPath);
-      }),
-    ).then((parsedRequirements) => parsedRequirements.flat());
+    const dependenciesFromRequirementsTxt = (
+      await Promise.all(
+        config.requirementsTxtFilePaths.map(readRequirementsTxt),
+      )
+    ).flat();
     logger.info(
       "Dependencies from `requirements.txt` files: %j",
       dependenciesFromRequirementsTxt,
@@ -417,43 +182,30 @@ yargs(hideBin(process.argv))
     ]);
     logger.info("Validated dependency list: %j", dependencies);
 
-    await copyBuildDirectory({ copyTo: destDir, keepOld: args.keepOldBuild });
+    // Step 1: lay down the desktop SPA shell into destDir.
+    await copyShellDirectory({
+      sourceDir: path.resolve(__dirname, pathFromScriptToBuild),
+      destDir,
+      keepOld: args.keepOldBuild,
+    });
 
-    const usedPrebuiltPackages = await saveUsedPrebuiltPackages({
+    // Step 2: vendor Pyodide deps + copy app files + write prebuilt-packages.txt.
+    const wheelsDir = path.join(__dirname, pathFromScriptToWheels);
+    await packageApp({
+      destDir,
+      sourceDir: projectDir,
+      files: config.files,
+      entrypoint: config.entrypoint,
+      dependencies,
+      localWheelPaths: [
+        path.join(wheelsDir, stliteLibWheelFileName),
+        path.join(wheelsDir, streamlitWheelFileName),
+      ],
       pyodideSource: args.pyodideSource,
-      pyodideRuntimeDir: path.resolve(destDir, "./pyodide"),
-      requirements: dependencies,
+      logger: winstonAdapter,
     });
-    logger.info(
-      "The prebuilt packages loaded for the given requirements: %j",
-      usedPrebuiltPackages,
-    );
 
-    const destAppDir = path.resolve(destDir, "./app_files"); // This path will be loaded in the `readStreamlitAppDirectory` handler in electron/main.ts.
-    await copyAppDirectory({
-      cwd: projectDir,
-      filePathPatterns: config.files,
-      destAppDir,
-    });
-    assertAppDirContainsEntrypoint(destAppDir, config.entrypoint);
-
-    await createSitePackagesSnapshot({
-      requirements: dependencies,
-      usedPrebuiltPackages,
-      pyodideSource: args.pyodideSource,
-      pyodideRuntimeDir: path.resolve(destDir, "./pyodide"),
-      saveTo: path.resolve(destDir, "./site-packages-snapshot.tar.gz"), // This path will be loaded in the `readSitePackagesSnapshot` handler in electron/main.ts.
-    });
-    // These prebuilt packages will be vendored in the build artifact by `downloadPrebuiltPackageWheels()`
-    // and the package names will be saved in the `./prebuilt-packages.txt` file
-    // so that they will be read and passed to `pyodide.loadPackage()` at runtime to install them from the vendored files.
-    // While the packages downloaded from PyPI at build time will have been shipped in the site-packages snapshot by `createSitePackagesSnapshot()`,
-    // the prebuilt packages must be installed at runtime by `pyodide.loadPackage()` or `micropip.install()`
-    // to avoid problems such as https://github.com/whitphx/stlite/issues/564.
-    await writePrebuiltPackagesTxt(
-      path.resolve(destDir, "./prebuilt-packages.txt"), // This path will be loaded in the `readRequirements` handler in electron/main.ts.
-      usedPrebuiltPackages,
-    );
+    // Step 3: desktop-specific manifest.
     await dumpManifest({
       packageJsonStliteDesktopField: packageJson.stlite?.desktop,
       manifestFilePath: path.resolve(destDir, "./stlite-manifest.json"),
