--- packages/desktop/bin-src/dump_artifacts/manifest.spec.ts
@@ -1,51 +1,0 @@
-import { describe, it, expect } from "vitest";
-import { coerceDesktopAppManifest } from "./manifest";
-
-describe("coerceDesktopAppManifest", () => {
-  it("should require the entrypoint field and coerce the remaining fields", () => {
-    expect(
-      coerceDesktopAppManifest({
-        entrypoint: "foo.py",
-      }),
-    ).toEqual({
-      entrypoint: "foo.py",
-      embed: false,
-      nodeJsWorker: false,
-      appMenu: true,
-    });
-  });
-
-  it("should allow appMenu to be set to false", () => {
-    expect(
-      coerceDesktopAppManifest({
-        entrypoint: "foo.py",
-        appMenu: false,
-      }),
-    ).toEqual({
-      entrypoint: "foo.py",
-      embed: false,
-      nodeJsWorker: false,
-      appMenu: false,
-    });
-  });
-
-  it("should throw an error when `idbfsMountpoints` is set with `nodeJsWorker` is true", () => {
-    expect(() =>
-      coerceDesktopAppManifest({
-        entrypoint: "foo.py",
-        nodeJsWorker: true,
-        idbfsMountpoints: ["foo"],
-      }),
-    ).toThrowError();
-  });
-
-  it("should throw an error when `nodefsMountpoints` is set with `nodeJsWorker` is false", () => {
-    expect(() =>
-      coerceDesktopAppManifest({
-        entrypoint: "foo.py",
-        nodeJsWorker: false,
-        nodefsMountpoints: { foo: "bar" },
-      }),
-    ).toThrowError();
-  });
-});
