+++ packages/desktop/scripts/build_bin.js
@@ -22,9 +22,8 @@ esbuild
     platform: "node",
     format: "esm",
     external: [
-      "pyodide", // The `pyodide` package must be installed at runtime for the included Wasm files, so there is no reason to bundle it here.
-      "fs-extra", // `fs-extra` and `yargs` will be installed at runtime anyway as the dependencies of `electron-builder`, so we don't have to bundle them here.
-      "yargs", // Same as above, `fs-extra`.
+      "yargs", // Installed at runtime as a dependency of electron-builder; no need to bundle.
+      "@stlite/app-packager", // Lives in its own npm package; transitively brings in pyodide, fs-extra, glob.
     ],
     define: {
       "process.env.PATH_FROM_SCRIPT_TO_BUILD": JSON.stringify("../build"),
